requirements.txt
gdown==4.6.4
numpy==1.23.5
Pillow==9.4.0
timm==0.6.12
torch==2.0.0
torchvision==0.15.0
